+++ projects/ng-dynamic-forms/core/src/lib/model/form-group/dynamic-form-group.model.spec.ts
@@ -21,7 +21,7 @@ describe("DynamicFormGroupModel test suite", () => {
     it("should initialize correctly", () => {
 
         expect(model.id).toEqual(config.id);
-        expect(model.group.length).toBe(1);
+        expect(model.group.length).toBe(config.group.length);
         expect(model.size()).toBe(model.group.length);
         expect(model.legend).toBeNull();
         expect(model.type).toEqual(DYNAMIC_FORM_CONTROL_TYPE_GROUP);